dev/capture: Limit captures to enabled contexts
  Some test scenarios can be disabled for various reasons on specific
  platforms like windows (i.e. `win32`) or be run only with local,
  remote or stopped client actions.
  The capture helper should not be run when the current context does not
  match these criterias so we don't end up with unnecessary captures
  that can lead us to believe they should be used.
  Even in situations where the scenario is disabled temporarily (e.g. a
  known bug prevents it from passing), we should avoid those unnecessary
  captures as they can quickly become outdated.

+++ dev/capture/local.js
@@ -196,6 +196,13 @@ const runAndRecordFSEvents =
     : runAndRecordChokidarEvents
 
 const captureScenario = (scenario /*: Scenario & {path: string} */) => {
+  if (
+    (scenario.platforms && !scenario.platforms.includes(process.platform)) ||
+    (scenario.side && scenario.side !== 'local')
+  ) {
+    return
+  }
+
   return fse
     .emptyDir(syncPath)
     .then(() => fse.emptyDir(outsidePath))
